fix hash to include measurement name
