+++ data/measurement.go
@@ -1,0 +1,27 @@
+package data
+
+import "hash/fnv"
+
+//TODO make this a method off of Measurment
+
+func calculateSeriesID(measurementName string, tags Tags) uint64 {
+	// <measurementName>|<tagKey>|<tagKey>|<tagValue>|<tagValue>
+	// cpu|host|servera
+	encodedTags := tags.Marshal()
+	size := len(measurementName) + len(encodedTags)
+	if len(encodedTags) > 0 {
+		size++
+	}
+	b := make([]byte, 0, size)
+	b = append(b, measurementName...)
+	if len(encodedTags) > 0 {
+		b = append(b, '|')
+	}
+	b = append(b, encodedTags...)
+	// TODO pick a better hashing that guarantees uniqueness
+	// TODO create a cash for faster lookup
+	h := fnv.New64a()
+	h.Write(b)
+	sum := h.Sum64()
+	return sum
+}
