+++ data/tag.go
@@ -1,0 +1,39 @@
+package data
+
+import "sort"
+
+type Tags map[string]string
+
+func (t Tags) Marshal() []byte {
+	// Empty maps marshal to empty bytes.
+	if len(t) == 0 {
+		return nil
+	}
+
+	// Extract keys and determine final size.
+	sz := (len(t) * 2) - 1 // separators
+	keys := make([]string, 0, len(t))
+	for k, v := range t {
+		keys = append(keys, k)
+		sz += len(k) + len(v)
+	}
+	sort.Strings(keys)
+
+	// Generate marshaled bytes.
+	b := make([]byte, sz)
+	buf := b
+	for _, k := range keys {
+		copy(buf, k)
+		buf[len(k)] = '|'
+		buf = buf[len(k)+1:]
+	}
+	for i, k := range keys {
+		v := t[k]
+		copy(buf, v)
+		if i < len(keys)-1 {
+			buf[len(v)] = '|'
+			buf = buf[len(v)+1:]
+		}
+	}
+	return b
+}
